requirements.txt
anyio==4.9.0
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
overrides==7.7.0
packaging==25.0
pip==25.1
platformdirs==4.3.8
prometheus_client==0.22.1
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack_data==0.6.3
tomli==2.2.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_utils==0.1.0
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
zipp==3.23.0
absl-py==2.3.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
fastapi==0.129.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.73.1
h5py==3.14.0
huggingface-hub==0.33.1
imbalanced-learn==0.14.1
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
markdown==3.8.2
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.6.2
multiprocess==0.70.16
namex==0.1.0
networkx==3.5
numpy==1.26.4
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.16.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf==4.25.8
pyarrow==20.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.2.3
python-dotenv==1.2.1
regex==2024.11.6
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sklearn-compat==0.1.5
sqlalchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.53.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.2.0
xxhash==3.5.0
yarl==1.20.1
pytest==9.0.2
